requirements.txt
Flask
openai
google-generativeai
pypdfium2
Pillow